refactor(app): rename user/ directory to orchestrator/
After SessionRunner extraction, the app/user/ submodules are all
"impl User" blocks split by orchestration concern (lifecycle,
inbound, freeze, consensus, steward, ...). The "user" name predates
the split; "orchestrator" reflects the actual scope.
